+++ Ray.Core/EventSourcing/IEventStorage.cs
@@ -6,8 +6,8 @@ namespace Ray.Core.EventSourcing
 {
     public interface IEventStorage<K>
     {
-        Task<List<EventInfo<K>>> GetListAsync(K stateId, UInt32 startVersion, UInt32 endVersion, DateTime? startTime = null);
-        Task<List<EventInfo<K>>> GetListAsync(K stateId, string typeCode, UInt32 startVersion, UInt32 endVersion, DateTime? startTime = null);
+        Task<List<EventInfo<K>>> GetListAsync(K stateId, Int64 startVersion, Int64 endVersion, DateTime? startTime = null);
+        Task<List<EventInfo<K>>> GetListAsync(K stateId, string typeCode, Int64 startVersion, Int64 endVersion, DateTime? startTime = null);
         Task<bool> SaveAsync<T>(T data, byte[] bytes, string uniqueId = null) where T : IEventBase<K>;
         Task CompleteAsync<T>(T data) where T : IEventBase<K>;
     }
