+++ Ray.Core/EventSourcing/IState.cs
@@ -8,11 +8,11 @@ namespace Ray.Core.EventSourcing
         /// <summary>
         /// 状态的版本号
         /// </summary>
-        UInt32 Version { get; set; }
+        Int64 Version { get; set; }
         /// <summary>
         /// 正在处理中的Version
         /// </summary>
-        UInt32 DoingVersion { get; set; }
+        Int64 DoingVersion { get; set; }
         /// <summary>
         /// 状态版本号对应的Event时间
         /// </summary>
